pyproject.toml
[project]
name = "capstone-project"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
    "matplotlib>=3.10.7",
    "mediacloud>=4.5.0",
    "networkx>=3.6",
    "pandas>=2.3.3",
    "python-dotenv>=1.2.1",
    "scikit-learn>=1.7.2",
    "seaborn>=0.13.2",
    "sentence-transformers>=5.1.2",
    "textblob>=0.19.0",
    "tqdm>=4.67.1",
    "vadersentiment>=3.3.2",
    "wordcloud>=1.9.4",
]
